Refactor repository and improve project structure
Renamed and refactored key files for consistency and alignment with naming conventions (e.g., `WriterSql` to `WriterSQL`). Adjusted package imports and reorganized the `repository` structure for better modularity. Updated the configuration and dependency list to ensure compatibility with the changes.
