+++ app/user/repository/sql/writer_sql.go
@@ -1,26 +1,27 @@
 package sql
 
 import (
-	"gorm.io/gorm"
 	entity "user-management/app/user"
+
+	"gorm.io/gorm"
 )
 
-type WriterSql struct {
+type WriterSQL struct {
 	DB *gorm.DB
 }
 
-func NewWriterSQL(db *gorm.DB) *WriterSql {
-	return &WriterSql{DB: db}
+func NewWriterSQL(db *gorm.DB) *WriterSQL {
+	return &WriterSQL{DB: db}
 }
 
-func (r *WriterSql) Create(user *entity.User) error {
+func (r *WriterSQL) Create(user *entity.User) error {
 	return r.DB.Create(user).Error
 }
 
-func (r *WriterSql) Update(user *entity.User) error {
+func (r *WriterSQL) Update(user *entity.User) error {
 	return r.DB.Save(user).Error
 }
 
-func (r *WriterSql) Delete(id int) error {
+func (r *WriterSQL) Delete(id int) error {
 	return r.DB.Delete(&entity.User{}, id).Error
 }
